requirements.txt
absl-py==1.4.0
chex==0.1.82
cloudpickle==2.2.1
contourpy==1.1.1
cycler==0.11.0
dm-env==1.6
dm-tree==0.1.8
filelock==3.9.0
fonttools==4.42.1
fsspec==2023.4.0
gym==0.26.2
gym-notices==0.0.8
jax==0.4.16
jaxlib==0.4.16
Jinja2==3.1.2
jumanji==0.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.8.0
ml-dtypes==0.2.0
mpmath==1.2.1
networkx==3.0rc1
numpy==1.26.0
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.11.2
six==1.16.0
sympy==1.11.1
tensordict-nightly==2023.9.19
toolz==0.12.0
torch==2.2.0.dev20230919+cpu
torchrl-nightly==2023.9.19
typing_extensions==4.4.0
